Split up the backend to clean up code
